+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Fixed
 
+- Nomen (was "castor") batch load was not assigning parent correctly [#3409]
 - Source and People API endpoints don't try to authenticate [#3407]
 - Date range in filter Collection Object not being applied [#3418]
 - Year import in NOMEN (was "castor") import [#3411]
@@ -41,6 +42,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Changing the number of items per page or the page shows loan items that are not from the current loan in New/edit loan task [#3413]
 - Sort by name gives an wrong order in filter nomenclature task
 
+[#3409]: https://github.com/SpeciesFileGroup/taxonworks/issues/3409
 [#3407]: https://github.com/SpeciesFileGroup/taxonworks/issues/3407
 [#3418]: https://github.com/SpeciesFileGroup/taxonworks/issues/3418
 [#3415]: https://github.com/SpeciesFileGroup/taxonworks/issues/3415